+++ NEWS.md
@@ -8,3 +8,4 @@
     - The Typst compiler prints error messages to `stderr`
     - If not `ignorestatus`, a Typst compiler error will throw a Julia `TypstError`
 - Increase coverage of the `Cmd` interface implementation for `TypstCommand`
+- Patch incorrect output from an assumption in `repr(::MIME, ::TypstString)`